CRM-10 Authentication

+++ powerful-crm.API/Controllers/LeadController.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
@@ -8,25 +9,31 @@ using powerful_crm.API.Models.OutputModels;
 using powerful_crm.Business;
 using powerful_crm.Core;
 using powerful_crm.Core.CustomExceptions;
+using powerful_crm.Core.Enums;
 using powerful_crm.Core.Models;
 using powerful_crm.Core.Settings;
 using RestSharp;
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Security.Claims;
 using System.Text.Json;
 
 namespace powerful_crm.API.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     public class LeadController : ControllerBase
     {
         private ILeadService _leadService;
+        private ICityService _cityService;
         private IMapper _mapper;
         private RestClient _client;
 
-        public LeadController(IOptions<AppSettings> options,IMapper mapper, ILeadService leadService)
+        public LeadController(IOptions<AppSettings> options,IMapper mapper, ILeadService leadService, ICityService cityService)
         {
             _leadService = leadService;
+            _cityService = cityService;
             _mapper = mapper;
             _client = new RestClient(options.Value.TSTORE_URL);
         }
@@ -36,6 +43,7 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(LeadOutputModel), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [AllowAnonymous]
         [HttpPost]
         public ActionResult<LeadOutputModel> AddLead([FromBody] LeadInputModel inputModel)
         {
@@ -43,9 +51,9 @@ namespace powerful_crm.API.Controllers
             {
                 throw new CustomValidationException(ModelState);
             }
-            if (_leadService.GetCityById(inputModel.CityId) == null)
+            if (_cityService.GetCityById(inputModel.CityId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_CITYNOTFOUND, inputModel.CityId));
+                return NotFound(string.Format(Constants.ERROR_CITY_NOT_FOUND, inputModel.CityId));
             }
             var dto = _mapper.Map<LeadDto>(inputModel);
             var addedLeadId = _leadService.AddLead(dto);
@@ -61,17 +69,20 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPut("{leadId}/change-password")]
         public ActionResult ChangePassword(int leadId, [FromBody]ChangePasswordInputModel inputModel)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             if (!ModelState.IsValid)
-            {
                 throw new CustomValidationException(ModelState);
-            }
+
             if (_leadService.GetLeadById(leadId) == null)
-            {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
-            }
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
+
             _leadService.ChangePassword(leadId, inputModel.OldPassword, inputModel.NewPassword);
             return NoContent();
         }
@@ -80,13 +91,14 @@ namespace powerful_crm.API.Controllers
         /// <returns>Info about lead</returns>
         [ProducesResponseType(typeof(LeadOutputModel), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpGet("{leadId}")]
         public ActionResult<LeadOutputModel> GetLead(int leadId)
         {
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
 
             var outputModel = _mapper.Map<LeadOutputModel>(lead);
@@ -98,6 +110,7 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(LeadOutputModel), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost("search")]
         public ActionResult<List<LeadOutputModel>> SearchLeads ([FromBody] SearchLeadInputModel inputModel)
         {
@@ -122,9 +135,14 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(LeadOutputModel), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPut("{leadId}")]
         public ActionResult<LeadOutputModel> UpdateLead(int leadId, [FromBody] UpdateLeadInputModel inputModel)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             if (!ModelState.IsValid)
             {
                 throw new CustomValidationException(ModelState);
@@ -132,11 +150,11 @@ namespace powerful_crm.API.Controllers
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
-            if ( _leadService.GetCityById(inputModel.CityId) == null)
+            if ( _cityService.GetCityById(inputModel.CityId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_CITYNOTFOUND, inputModel.CityId));
+                return NotFound(string.Format(Constants.ERROR_CITY_NOT_FOUND, inputModel.CityId));
             }
             var dto = _mapper.Map<LeadDto>(inputModel);
             _leadService.UpdateLead(leadId, dto);
@@ -151,17 +169,22 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(List<LeadOutputModel>), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpDelete("{leadId}")]
         public ActionResult<LeadOutputModel> DeleteLead(int leadId)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
             if (lead.IsDeleted == true)
             {
-                return BadRequest(string.Format(Constants.ERROR_LEADALREADYDELETED, leadId));
+                return BadRequest(string.Format(Constants.ERROR_LEAD_ALREADY_DELETED, leadId));
             }
             _leadService.DeleteLead(leadId);
             var dto = _mapper.Map<LeadOutputModel>(_leadService.GetLeadById(leadId));
@@ -174,77 +197,49 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(List<LeadOutputModel>), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPut("{leadId}/recover")]
         public ActionResult<LeadOutputModel> RecoverLead(int leadId)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
             if (lead.IsDeleted == false)
             {
-                return BadRequest(string.Format(Constants.ERROR_LEADNOTDELETED, leadId));
+                return BadRequest(string.Format(Constants.ERROR_LEAD_NOT_DELETED, leadId));
             }
             _leadService.RecoverLead(leadId);
             var dto = _mapper.Map<LeadOutputModel>(_leadService.GetLeadById(leadId));
             return Ok(dto);
         }
 
-        /// <summary>Creates new city</summary>
-        /// <param name="city">Information about new city</param>
-        /// <returns>Info about created city</returns>
-        [ProducesResponseType(typeof(CityOutputModel), StatusCodes.Status200OK)]
-        [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
-        [HttpPost("city")]
-        public ActionResult<CityOutputModel> AddCity([FromBody] CityInputModel city)
-        {
-            if (!ModelState.IsValid)
-            {
-                throw new CustomValidationException(ModelState);
-            }
-            var dto = _mapper.Map<CityDto>(city);
-            var addedCityId = _leadService.AddCity(dto);
-            var outputModel = _mapper.Map<CityOutputModel>(_leadService.GetCityById(addedCityId));
-            return Ok(outputModel);
-        }
-
-        /// <summary>Deletes the city</summary>
-        /// <param name="cityId">Id of the city to delete</param>
-        /// <returns>NoContent result</returns>
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
-        [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(typeof(string), StatusCodes.Status409Conflict)]
-        [HttpDelete("city/{id}")]
-        public ActionResult<LeadOutputModel> DeleteCity(int cityId)
-        {
-            var city = _leadService.GetCityById(cityId);
-            if (city == null)
-            {
-                return NotFound(string.Format(Constants.ERROR_CITYNOTFOUND, cityId));
-            }
-            if (_leadService.DeleteLead(cityId) == 1)
-                return NoContent();
-            else
-                return Conflict(string.Format(Constants.ERROR_CITYHASDEPENDENCIES, cityId));
-        }
-
+       
         /// <summary>Gets lead balance</summary>
         /// <param name="leadId">Id of lead</param>
         /// <returns>Info about balance</returns>
         [ProducesResponseType(typeof(List<BalanceOutputModel>), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpGet("{leadId}/balance")]
         public ActionResult<List<BalanceOutputModel>> GetBalanceByLeadId(int leadId)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
 
-            var request = new RestRequest(string.Format(Constants.API_GETBALANCE, leadId), Method.GET);
+            var request = new RestRequest(string.Format(Constants.API_GET_BALANCE, leadId), Method.GET);
             var queryResult = _client.Execute<List<BalanceInputModel>>(request).Data;
 
             var result = _mapper.Map<List<BalanceOutputModel>>(queryResult);
@@ -256,16 +251,21 @@ namespace powerful_crm.API.Controllers
         /// <returns>Info about transactions</returns>
         [ProducesResponseType(typeof(List<TransactionOutputModel>), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpGet("{leadId}/transactions")]
         public ActionResult<List<TransactionOutputModel>> GetTransactionsByLeadId(int leadId)
         {
+            if (!CheckIfUserIsAllowed(leadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             var lead = _leadService.GetLeadById(leadId);
             if (lead == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, leadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, leadId));
             }
 
-            var request = new RestRequest(string.Format(Constants.API_GETTRANSACTION, leadId), Method.GET);
+            var request = new RestRequest(string.Format(Constants.API_GET_TRANSACTION, leadId), Method.GET);
             var queryResult = _client.Execute<string>(request).Data;
 
             return Ok(queryResult);
@@ -277,14 +277,19 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(int), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost("deposit")]
         public ActionResult<int> AddDeposit([FromBody] TransactionInputModel inputModel)
         {
+            if (!CheckIfUserIsAllowed(inputModel.LeadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             if (!ModelState.IsValid)
                 throw new CustomValidationException(ModelState);
             if (_leadService.GetLeadById(inputModel.LeadId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, inputModel.LeadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, inputModel.LeadId));
             }
             var middle = _mapper.Map<TransactionMiddleModel>(inputModel);
             var request = new RestRequest(Constants.API_DEPOSIT, Method.POST);
@@ -299,14 +304,19 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(int), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost("withdraw")]
         public ActionResult<int> AddWithdraw([FromBody] TransactionInputModel inputModel)
         {
+           if (!CheckIfUserIsAllowed(inputModel.LeadId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             if (!ModelState.IsValid)
                 throw new CustomValidationException(ModelState);
             if (_leadService.GetLeadById(inputModel.LeadId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, inputModel.LeadId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, inputModel.LeadId));
             }
             var middle = _mapper.Map<TransactionMiddleModel>(inputModel);
             var request = new RestRequest(Constants.API_WITHDRAW, Method.POST);
@@ -321,18 +331,24 @@ namespace powerful_crm.API.Controllers
         [ProducesResponseType(typeof(int), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
         [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost("transfer")]
         public ActionResult<int> AddTransfer([FromBody] TransferInputModel inputModel)
         {
+            if (!CheckIfUserIsAllowed(inputModel.SenderId))
+                throw new ForbidException(Constants.ERROR_NOT_ALLOWED_ACTIONS_WITH_OTHER_LEAD);
+
             if (!ModelState.IsValid)
                 throw new CustomValidationException(ModelState);
+
             if (_leadService.GetLeadById(inputModel.RecipientId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, inputModel.RecipientId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, inputModel.RecipientId));
             }
             if (_leadService.GetLeadById(inputModel.SenderId) == null)
             {
-                return NotFound(string.Format(Constants.ERROR_LEADNOTFOUND, inputModel.SenderId));
+                return NotFound(string.Format(Constants.ERROR_LEAD_NOT_FOUND_BY_ID, inputModel.SenderId));
             }
             var middle = _mapper.Map<TransferMiddleModel>(inputModel);
             var request = new RestRequest(Constants.API_TRANSFER, Method.POST);
@@ -341,5 +357,10 @@ namespace powerful_crm.API.Controllers
             return Ok(queryResult);
         }
 
+        private bool CheckIfUserIsAllowed(int leadId)
+        {
+            return leadId.ToString() == HttpContext.User.Claims.Where(t=>t.Type==ClaimTypes.NameIdentifier).FirstOrDefault().Value 
+                || HttpContext.User.IsInRole(Role.Administrator.ToString());
+        }
     }
 }
